Move classes to droppable
